+++ src/Mod.cpp
@@ -5,6 +5,10 @@ bool drift = true;
 bool jumpDash = true;
 bool doubleJumpDash = true;
 bool lightDash = true;
+bool stompBounce = true;
+bool stompBounceLightning = true;
+bool stompBounceLightningFix = true;
+bool legacyControls = false;
 
 EXPORT void Init()
 {
@@ -19,6 +23,10 @@ EXPORT void Init()
 	jumpDash = states["jumpDash"] == "True";
 	doubleJumpDash = states["doubleJumpDash"] == "True";
 	lightDash = states["lightDash"] == "True";
+	stompBounce = states["stompBounce"] == "True";
+	stompBounceLightning = states["stompBounceLightning"] == "True";
+	stompBounceLightningFix = states["stompBounceLightningFix"] == "True";
+	legacyControls = states["legacyControls"] == "True";
 }
 
 EXPORT void PostInit()