+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# V1.1.0
+
+- Restored the Stomp Bounce
+- Restored sound for Stomping and Light Speed Dash
+- Changed Drifting from LT to B on controller
+- Fixed not being able to Light Speed Dash from a jump
+- Added a Legacy Controls option for anyone using the in-game Legacy Controls option
+
 # V1.0.1
 
 - Fixes broken input when disabling Drift